Claude/add local run flutter login 013688v b so tt may l byi f sto x

+++ backend/src/main/java/com/dutyout/application/service/AutoScheduleService.java
@@ -84,7 +84,7 @@ public class AutoScheduleService {
         // 3. 기존 스케줄 확인 및 삭제 (오늘 날짜)
         // unique 제약 위반 방지를 위해 기존 스케줄을 먼저 삭제
         LocalDate today = LocalDate.now();
-        long deletedCount = dailyScheduleRepository.deleteByBabyIdAndScheduleDate(babyId, today);
+        int deletedCount = dailyScheduleRepository.deleteByBabyIdAndScheduleDate(babyId, today);
         log.info("기존 스케줄 삭제 완료 - 삭제 건수: {}", deletedCount);
 
         // 4. 스케줄 아이템 생성
